deep_learning/kunskapskontroll_1/requirements.txt
tensorflow==2.17.1           
tensorflow-io-gcs-filesystem==0.37.1
numpy==1.26.4                
pandas==2.0.3
scipy==1.13.1
scikit-image==0.21.0
opencv-python==4.8.0.76
matplotlib==3.7.1
seaborn==0.12.2
streamlit==1.25.0
av==10.0.0
h5py==3.13.0
Pillow==9.5.0
networkx==3.2.1
PyWavelets==1.6.0
mtcnn==0.1.1                 
